Extract shared livestream channel logic into LivestreamChannelHandler
The handler owns the in-memory channel and messages state, pause/skip
bookkeeping, typing cleanup timers, and channel event processing that
will be shared between LivestreamChannelController and the upcoming
LivestreamChat state-layer component. State changes are surfaced
through a Handlers struct so owners can wire them to their preferred
callback mechanism.

LivestreamChannelController becomes a thin facade that delegates to
the handler while preserving its public delegate-based interface.
